pyproject.toml
[project]
name = "news-summarizer-bot"
version = "0.1.0"
description = "Slack bot that summarizes news articles posted in a channel"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "anthropic>=0.40.0",
    "trafilatura>=1.12.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
]
